Add checks for pre-existing files to downloader (#67)
* Check if file exists before download - downloader

* Update documentation
